+++ lib/presentation/pages/admin_profile/user_transactions_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:meta_app/presentation/widgets/profile_header/header.dart';
+import '../../widgets/profile_header/profile_header.dart';
 
 class UserTransactionsPage extends StatelessWidget {
   const UserTransactionsPage({super.key});
